Removing the GTK3 subplot tool and replacing it with generic one
